fix(security): Ensure project permissions are checked for code mappings endpoints (#64580)
Addresses https://github.com/getsentry/sentry/issues/55871

Will now use the request context for projects to enforce membership for
organizations without open enrollment in teams. Added tests to double
check this behaviour.

+++ src/sentry/api/endpoints/organization_code_mapping_details.py
@@ -59,6 +59,8 @@ class OrganizationCodeMappingDetailsEndpoint(OrganizationEndpoint, OrganizationI
         :param string default_branch:
         :auth: required
         """
+        if not request.access.has_project_access(config.project):
+            return self.respond(status=status.HTTP_403_FORBIDDEN)
 
         try:
             # We expect there to exist an org_integration
@@ -92,6 +94,10 @@ class OrganizationCodeMappingDetailsEndpoint(OrganizationEndpoint, OrganizationI
 
         :auth: required
         """
+
+        if not request.access.has_project_access(config.project):
+            return self.respond(status=status.HTTP_403_FORBIDDEN)
+
         try:
             config.delete()
             return self.respond(status=status.HTTP_204_NO_CONTENT)
